implement review comment to follow Otavio's preferences for packages
